moving test to root test dir
